+++ CHANGELOG.md
@@ -349,3 +349,4 @@
 - 2024-03-28: Rate users, discovery, follow up, task, +2
 - 2024-03-29: Retry users, mutate, contacts, task, +5
 - 2024-03-29: Init icp, sender, contacts, users, +2
+- 2024-03-30: Rate README, follow up, mutate, sender, +3